requirements.txt
aiogram==3.7.0
pandas==2.2.2
openpyxl==3.1.5
xlrd==1.2.0
Pillow==10.4.0
Flask==3.0.3
